+++ deep_seq_batch/generate_sub_clusters.contnued_from_06.R
@@ -21,13 +21,14 @@ num.dim.cluster <- 30
 cluster.resolution <- 0.5
 use.sctransform <- TRUE
 vars.to.regress <- c("percent.mt")
-
+my_random_seed <- 42
 outdir <- "/media/hieunguyen/HD01/outdir/CRC1382/SBharadwaj_20240318"
-PROJECT <- "SBharadwaj_20240318_Sample_1_4_7_8"
+PROJECT <- "SBharadwaj_20240318_Sample_1_4_7_8_2_5"
 sub.cluster.idx <- "v0.1"
 # cont.sub.cluster.idx <- "Monocyte_Macrophages"
 
 for (cont.sub.cluster.idx in names(sub_clusters[[PROJECT]][[sub.cluster.idx]])){
+  print(sprintf("Working on sub cluster: %s", cont.sub.cluster.idx))
   path.to.main.input <- file.path(outdir, PROJECT)
   path.to.main.output <- file.path(path.to.main.input, "data_analysis")
   path.to.06.output <- file.path(path.to.main.output, "06_output", sub.cluster.idx)
@@ -38,21 +39,26 @@ for (cont.sub.cluster.idx in names(sub_clusters[[PROJECT]][[sub.cluster.idx]])){
   s.obj.raw <- readRDS(file.path(path.to.06.output, sprintf("Project_%s_%s.rds", PROJECT, sub.cluster.idx)))
   s.obj <- subset(s.obj.raw, cca.cluster.0.5 %in% sub_clusters[[PROJECT]][[sub.cluster.idx]][[cont.sub.cluster.idx]])
   
-  s.obj.no.integrated <- s.obj
-  s.obj.no.integrated <- DietSeurat(s.obj.no.integrated)
-  pca_reduction_name <- "RNA_PCA"
-  umap_reduction_name <- "RNA_UMAP"
-  
-  s.obj.no.integrated <- RunPCA(s.obj.no.integrated, npcs = num.PCA, verbose = FALSE, reduction.name=pca_reduction_name)
-  s.obj.no.integrated <- RunUMAP(s.obj.no.integrated, reduction = pca_reduction_name, 
-                   dims = 1:num.PC.used.in.UMAP, reduction.name=umap_reduction_name,
-                   seed.use = my_random_seed, umap.method = "uwot")
-  # clustering 
-  s.obj.no.integrated <- FindNeighbors(s.obj.no.integrated, reduction = pca_reduction_name, dims = 1:num.PC.used.in.Clustering)
-  s.obj.no.integrated <- FindClusters(s.obj.no.integrated, resolution = cluster.resolution, random.seed = 0)
-  dir.create(file.path(path.to.07.output, "s8_output"), showWarnings = FALSE, recursive = TRUE)
-  
-  saveRDS(s.obj.no.integrated, file.path(path.to.07.output, "s8_output", sprintf("%s_%s.noIntegration.rds", PROJECT, cont.sub.cluster.idx)))
+  if (file.exists(file.path(path.to.07.output, "s8_output", sprintf("%s_%s.noIntegration.rds", PROJECT, cont.sub.cluster.idx))) == FALSE){
+    s.obj.no.integrated <- s.obj
+    s.obj.no.integrated <- DietSeurat(s.obj.no.integrated)
+    DefaultAssay(s.obj.no.integrated) <- "RNA"
+    s.obj.no.integrated <- SCTransform(s.obj.no.integrated, vars.to.regress = vars.to.regress, verbose = FALSE)
+    pca_reduction_name <- "SCT_PCA"
+    umap_reduction_name <- "SCT_UMAP"
+    
+    s.obj.no.integrated <- RunPCA(s.obj.no.integrated, npcs = num.PCA, verbose = FALSE, reduction.name=pca_reduction_name)
+    s.obj.no.integrated <- RunUMAP(s.obj.no.integrated, reduction = pca_reduction_name, 
+                                   dims = 1:num.PC.used.in.UMAP, reduction.name=umap_reduction_name,
+                                   seed.use = my_random_seed, umap.method = "uwot")
+    # clustering 
+    s.obj.no.integrated <- FindNeighbors(s.obj.no.integrated, reduction = pca_reduction_name, dims = 1:num.PC.used.in.Clustering)
+    s.obj.no.integrated <- FindClusters(s.obj.no.integrated, resolution = cluster.resolution, random.seed = 0)
+    
+    dir.create(file.path(path.to.07.output, "s8_output"), showWarnings = FALSE, recursive = TRUE)
+    
+    saveRDS(s.obj.no.integrated, file.path(path.to.07.output, "s8_output", sprintf("%s_%s.noIntegration.rds", PROJECT, cont.sub.cluster.idx)))
+  } 
   
   ##### re-integration
   if (file.exists(file.path(path.to.07.output, "s8_output", sprintf("%s_%s.output.s8.rds", PROJECT, cont.sub.cluster.idx))) == FALSE){
